+++ shipgen/FixedTargetGenerator.cxx
@@ -223,18 +223,21 @@ Bool_t FixedTargetGenerator::Init()
      LOG(fatal) << "EVTGENDATA environment variable not set";
    }
    TString DecayFile = TString(evtgendata) + "/DECAY.DEC";
+
    TString ParticleFile = TString(evtgendata) + "/evt.pdl";
    std::cout << "Using $EVTGENDATA " << evtgendata << std::endl;
    EvtAbsRadCorr *fsrPtrIn = 0;
    EvtExternalGenList *extPtr = new EvtExternalGenList();
    std::list<EvtDecayBase*> models = extPtr->getListOfModels();
+#if PYTHIA_VERSION_INTEGER < 8315
  // Define the random number generator
    EvtRandomEngine* eng = new EvtSimpleRandomEngine();
    EvtRandom::setRandomEngine(eng);
    EvtGen *myEvtGenPtr = new EvtGen(DecayFile.Data(), ParticleFile.Data(),eng, fsrPtrIn, &models, 1, false);
+#endif
    TString UdecayFile    = getenv("FAIRSHIP");UdecayFile +="/gconfig/USERDECAY.DEC";
 #if PYTHIA_VERSION_INTEGER >= 8315
-   evtgenP = new Pythia8::EvtGenDecays(fPythiaP, DecayFile.Data(), ParticleFile.Data(),extPtr);
+   evtgenP = new Pythia8::EvtGenDecays(fPythiaP, DecayFile.Data(), ParticleFile.Data(), extPtr);
 #else
    evtgenP = new EvtGenDecays(fPythiaP, DecayFile.Data(), ParticleFile.Data(),myEvtGenPtr);
 #endif
@@ -242,7 +245,7 @@ Bool_t FixedTargetGenerator::Init()
    // use one instance of EvtGen, requires patch to Pythia8Plugins/EvtGen.h
    if (Option == "Primary"){
 #if PYTHIA_VERSION_INTEGER >= 8315
-    evtgenN = new Pythia8::EvtGenDecays(fPythiaN, DecayFile.Data(), ParticleFile.Data(),extPtr);
+    evtgenN = new Pythia8::EvtGenDecays(fPythiaN, DecayFile.Data(), "", extPtr);
 #else
     evtgenN = new EvtGenDecays(fPythiaN, DecayFile.Data(), ParticleFile.Data(),myEvtGenPtr);
 #endif